modify PlotVariables1d run method for multi config
