+++ columnflow/tasks/framework/plotting.py
@@ -518,3 +518,86 @@ class VariablePlotSettingMixin(
         dict_add_strict(params, "variable_settings", self.variable_settings)
 
         return params
+
+
+class PlotShiftMixin(AnalysisTask):
+    # TODO: details
+    shift = luigi.Parameter(
+        default="nominal",
+        description="the shift to plot; empty default",
+    )
+
+    @classmethod
+    def modify_param_values(cls, params):
+        """
+        When "config" and "shift" are set, this method evaluates them to set the global shift.
+        For that, it takes the shifts stored in the config instance and compares it with those
+        defined by this class.
+        """
+        params = super().modify_param_values(params)
+
+        # get params
+        config_insts = params.get("config_insts")
+        requested_shift = params.get("shift")
+
+        # require that the configs is set
+        if config_insts in (None, law.NO_STR, ()):
+            return params
+
+        # require that the shift is set and known
+        if requested_shift in (None, law.NO_STR):
+            if cls.allow_empty_shift:
+                params["shift"] = law.NO_STR
+                return params
+
+            raise Exception(f"no shift found in params: {params}")
+
+        for config_inst in config_insts:
+            if requested_shift not in config_inst.shifts:
+                raise ValueError(f"shift {requested_shift} unknown to {config_inst}")
+
+        # determine the known shifts for this class
+        shifts, upstream_shifts = cls.get_known_shifts(config_insts[0], params)
+
+        # actual shift resolution: compare the requested shift to known ones
+        # local_shift -> the requested shift if implemented by the task itself, else nominal
+        # shift       -> the requested shift if implemented by this task
+        #                or an upsteam task (== global shift), else nominal
+
+        if requested_shift in shifts:
+            params["shift"] = requested_shift
+        elif requested_shift in upstream_shifts:
+            params["shift"] = requested_shift
+        else:
+            params["shift"] = "nominal"
+
+        # store references
+        params["shift_inst"] = config_insts[0].get_shift(params["shift"])
+
+        return params
+
+    @classmethod
+    def resolve_param_values(cls, params: dict) -> dict:
+        params = super().resolve_param_values(params)
+
+        # set default shift
+        if params.get("shift") in (None, law.NO_STR):
+            params["shift"] = "nominal"
+
+        return params
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        # store references to the shift instances
+        self.shift_inst = None
+        self.global_shift_inst = None
+
+        if self.shift not in (None, law.NO_STR):
+            self.shift_inst = self.config_insts[0].get_shift(self.shift)
+
+    def store_parts(self):
+        parts = super().store_parts()
+        # add the shift name
+        parts.insert_after("analysis", "shift", self.shift)
+        return parts
